+++ Labyrinth/src/Labyrinth/Renderer/APIs/OpenGL/OpenGLTexture.h
@@ -9,18 +9,18 @@ namespace Labyrinth {
 	class LAB_API OpenGLTexture2D : public Texture2D
 	{
 	public:
-		OpenGLTexture2D(size_t width, size_t height);
+		OpenGLTexture2D(i32 width, i32 height);
 		OpenGLTexture2D(const std::string& path);
 		virtual ~OpenGLTexture2D();
 
 		virtual const std::string& getPath() const override { return mPath; }
-		virtual size_t getWidth() const override { return mWidth; }
-		virtual size_t getHeight() const override { return mHeight; }
-		virtual uint32_t getRendererID() const override { return mRendererID; }
+		virtual i32 getWidth() const override { return mWidth; }
+		virtual i32 getHeight() const override { return mHeight; }
+		virtual u32 getRendererID() const override { return mRendererID; }
 
-		virtual void setData(void* data, size_t size) override;
+		virtual void setData(void* data, i32 size) override;
 
-		virtual void bind(uint32_t slot = 0) const override;
+		virtual void bind(u32 slot = 0) const override;
 
 		virtual bool operator==(const Texture& other) const override
 		{
@@ -29,8 +29,8 @@ namespace Labyrinth {
 
 	private:
 		std::string mPath;
-		size_t mWidth, mHeight;
-		uint32_t mRendererID;
+		i32 mWidth, mHeight;
+		u32 mRendererID;
 		GLenum mInternalFormat, mDataFormat;
 	};
