+++ Labyrinth/src/Labyrinth/Renderer/SubTexture.cpp
@@ -10,16 +10,16 @@ namespace Labyrinth {
 	Texture2DSheet::Texture2DSheet(const Ref<Texture2D>& spriteSheet, const glm::vec2& tileSize, const std::string& name)
 		: mTexture(spriteSheet), mTileSize(tileSize), mName(name)
 	{
-		mTileCountX = (uint32_t)(round(mTexture->getWidth() / tileSize.x));
-		mTileCountY = (uint32_t)(round(mTexture->getHeight() / tileSize.y));
+		mTileCountX = (u32)(round(mTexture->getWidth() / tileSize.x));
+		mTileCountY = (u32)(round(mTexture->getHeight() / tileSize.y));
 		mSubTextures.reserve(mTileCountX * mTileCountY);
 	}
 
 	Texture2DSheet::Texture2DSheet(const std::string& filepath, const glm::vec2& tileSize, const std::string& name)
 		: mTexture(Texture2D::Create(filepath)), mTileSize(tileSize), mName(name)
 	{
-		mTileCountX = (uint32_t)(round(mTexture->getWidth() / tileSize.x));
-		mTileCountY = (uint32_t)(round(mTexture->getHeight() / tileSize.y));
+		mTileCountX = (u32)(round(mTexture->getWidth() / tileSize.x));
+		mTileCountY = (u32)(round(mTexture->getHeight() / tileSize.y));
 		mSubTextures.reserve(mTileCountX * mTileCountY);
 	}
 
@@ -89,16 +89,16 @@ namespace Labyrinth {
 		mSubTextures.erase(name);
 	}
 
-	void Texture2DSheet::generateTileset(size_t startIndex)
+	void Texture2DSheet::generateTileset(usize startIndex)
 	{
 		mSubTextures.clear();
 
-		int count = startIndex;
-		for (uint32_t y = 0; y < mTileCountY; y++)
+		usize count = startIndex;
+		for (u32 y = 0; y < mTileCountY; y++)
 		{
-			for (uint32_t x = 0; x < mTileCountX; x++)
+			for (u32 x = 0; x < mTileCountX; x++)
 			{
-				createSubTex(std::to_string(count), { Cast<float>(x), Cast<float>(y) });
+				createSubTex(std::to_string(count), { Cast<f32>(x), Cast<f32>(y) });
 				count++;
 			}
 		}
@@ -145,7 +145,7 @@ namespace Labyrinth {
 	SubTexture2D::SubTexture2D(const Ref<Texture2DSheet> sheet, const glm::vec2 coords[4], const std::string& name)
 		: mSheet(sheet), mName(name)
 	{
-		for (size_t i = 0; i < 4; i++)
+		for (usize i = 0; i < 4; i++)
 			mTexCoords[i] = coords[i];
 	}
 
@@ -162,8 +162,8 @@ namespace Labyrinth {
 	}
 	Ref<SubTexture2D> SubTexture2D::Create(const Ref<Texture2DSheet>& tex, const glm::vec2& coords, const glm::vec2& spriteSize, const std::string& name)
 	{
-		float sheetWidth = Cast<float>(tex->getWidth());
-		float sheetHeight = Cast<float>(tex->getHeight());
+		f32 sheetWidth = Cast<f32>(tex->getWidth());
+		f32 sheetHeight = Cast<f32>(tex->getHeight());
 
 		//Calculate y coord as Sheet Height - y so that "coords" parameter can be given as position from top right
 		glm::vec2 min = { ((coords.x * tex->mTileSize.x) / sheetWidth), ((sheetHeight - (coords.y * tex->mTileSize.y)) / sheetHeight) };
